requirements.txt
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==4.2.23
django-redis==6.0.0
djangorestframework==3.16.0
Faker==37.4.2
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.51
python-dateutil==2.9.0.post0
psycopg2-binary>=2.9
redis==6.2.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
Django==4.2.23
django-celery-beat==2.8.1
django-redis==6.0.0
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
Faker==37.4.2
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
python-crontab==3.3.0
python-dateutil==2.9.0.post0
redis==6.2.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
